Allow integer postfix multipliers at command line
Data for arguments and options requiring an integer input can now be provided in the following ways:
- Multiplier postfix: e.g. 6.2k = 6200.
- Scientific notation: 6.2e3 = 6200 (is detected as scientific notation, parsed as floating-point, then converted back to integer; previously such input would not have worked properly).
Closes #215.

+++ lib/cmdline_option.h
@@ -133,7 +133,7 @@ namespace MR
             int def;
           } choices;
           struct {
-            int def, min, max;
+            int64_t def, min, max;
           } i;
           struct {
             default_type def, min, max;
@@ -194,7 +194,7 @@ namespace MR
         //! specifies that the argument should be an integer
         /*! if desired, a default value can be specified, along with a range of
          * allowed values. */
-        Argument& type_integer (int min = std::numeric_limits<int>::min(), int def = 0, int max = std::numeric_limits<int>::max()) {
+        Argument& type_integer (int64_t min = std::numeric_limits<int64_t>::min(), int64_t def = 0, int64_t max = std::numeric_limits<int64_t>::max()) {
           type = Integer;
           defaults.i.min = min;
           defaults.i.def = def;
